+++ src/num/alg/vector/vec/mod.rs
@@ -1,9 +1,0 @@
-// devela/src/num/alg/vector/vec/mod.rs
-//
-//!
-//
-
-mod impl_traits;
-mod methods;
-mod num;
-mod ops;